refactor: remove barrel index files, replace React namespace imports with named imports
