+++ webapp/client/apps/orchestration-cluster-webapp/src/operate/shared/EmptyMessage/EmptyMessage.tsx
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React from 'react';
+import {type FC} from 'react';
 import {Message, AdditionalInfo, Stack} from './styled';
 
 type Props = {
@@ -14,7 +14,7 @@ type Props = {
 	additionalInfo?: string;
 };
 
-const EmptyMessage: React.FC<Props> = ({message, additionalInfo, ...props}) => {
+const EmptyMessage: FC<Props> = ({message, additionalInfo, ...props}) => {
 	return (
 		<Stack {...props} gap={3}>
 			<Message>{message}</Message>
